fix: build on a plain Theos install
RemoteLog.h is a debugging header the author has installed locally, so
the repo did not compile for anyone else. It is now optional and falls
back to NSLog. Also casts the trait initialiser, which C++ will not
narrow implicitly.

+++ SNVersionTraits.h
@@ -53,7 +53,7 @@ static inline SNVersionTraits SNTraits(void) {
         BOOL isLegacy = (version > 0.0f && version <= 458.0f);
 
         traits = (SNVersionTraits){
-            .fieldTypeShift            = isLegacy ? 1 : 0,
+            .fieldTypeShift            = (NSUInteger)(isLegacy ? 1 : 0),
             .lightSpeedSymbolsInEngine = !isLegacy,
             .usesADTInfoInitialiser    = !isLegacy,
             .remapsMdsColorCodes       = (version == 458.0f),
